+++ packages/runtime/src/reactRouterDom.ts
@@ -1,0 +1,31 @@
+// https://github.com/ReactTraining/react-router/tree/master/packages/react-router-dom
+
+export {
+  // components
+  Link,
+  NavLink,
+  Prompt,
+  Redirect,
+  Route,
+  Switch,
+  // BrowserRouter,
+  // HashRouter,
+  // MemoryRouter,
+  // StaticRouter,
+
+  // static method
+  // export withrouter API from the create-app-shared
+  // withRouter,
+  matchPath,
+  generatePath,
+
+  // hooks
+  useHistory,
+  useLocation,
+  useParams,
+  useRouteMatch,
+
+  RouteComponentProps,
+  Router,
+  StaticRouter,
+} from 'react-router-dom';